Rename category icons to avoid overriding application icons
These new icons are designed to be used in gnome-control-center, but at least
one of them (preferences-system.svg) is used by as the application icon of
gnome-control-center itself as well as other applications such as
gnome-tweak-tool.

This change goes with a matching change to the gnome-menu package which appends
the "-directory" suffix to the Icon key of the Personal.directory,
System.directory and Hardware.directory files.