Don't draw backgrounds and borders on menu items
For some reason (probably debugging?) the CSS of the gtk themes was
requesting a border to be drawn around menu items.  They also had an
inherited background colour.  Gtk didn't used to draw these, but CSS
support has improved in 3.8 so the background colour and the border
become visible.

Drop those so that they don't get drawn in the new Gtk.