Adjust CSS to new Gtk menu padding regime
Gtk 3.6 featured a horizontal-padding property on GtkMenuBar with a
default value of 3.  This was added to the padding of all menuitems in
the bar.  Upstream has dropped this property, advising to use the CSS 
padding property directly.

Adjust to these changes by increasing the normal horizontal 'padding:' 
on our menuitems by 3px.  Retain backwards compatibility with older 
versions of Gtk by setting the horizontal-padding property to zero in 
these versions (to remove the default 3 pixels that used to be added, 
preventing us from having 6 pixels total).