Don't use GtkWidget-wide-separators
There are two code paths in gtk for drawing menu separators.  The one we were
using (which is able to draw fancy rectangles) uses css border properties.  In
gtk 3.8, menu items honor the border property, so we get a border around every
separator.

This patch makes it use the other code path, which draws a simple line.  It
also removes borders.