Fix Ambiance's gtk3 appearance for readonly Entry widgets
Inspired in upstream's fix https://bugzilla.gnome.org/show_bug.cgi?id=694831
the GtkEntry elements which are readonly (editable=false) should have a
different appearance. In this case, we use the background colour that is
used already for insensitive ones (this will not make them be confused
with insensitive ones, because the latter also have a grey colour for
the text, which doesn't happen for readonly widgets).

Fixes https://bugs.launchpad.net/ubuntu/+source/ubuntu-themes/+bug/1219205